+++ common/util.go
@@ -54,6 +54,7 @@ import (
 	"sync"
 	"sync/atomic"
 	"syscall"
+	"time"
 
 	"github.com/petermattis/goid"
 	"gopkg.in/ini.v1"
@@ -648,4 +649,50 @@ func PrettyOpenFlags(f int) string {
 // runtime.Stack calls.
 func GetGoroutineID() uint64 {
 	return (uint64)(goid.Get())
+}
+
+const layoutValidityDuration = 5 * time.Minute
+
+type LayoutRange struct {
+	Start    int64
+	End      int64
+	Endpoint string
+}
+
+type Layout struct {
+	LayoutRanges    []LayoutRange
+	LastFetchedTime atomic.Int64
+}
+
+func (l *Layout) Invalidate() {
+	l.LastFetchedTime.Store(0)
+}
+
+// TODO: Implement refreshing the layout if it is invalid.
+func (l *Layout) IsValid() bool {
+	lastFetched := time.Unix(l.LastFetchedTime.Load(), 0)
+	return time.Since(lastFetched) < layoutValidityDuration
+}
+
+// getIdealEndpoint returns the endpoint to download the blob at the given offset based on the layout.
+// It considers the first range containing the offset. If the chunk isn't fully available, the service
+// still returns correct data.
+func (l *Layout) GetIdealEndpoint(offset int64) string {
+	if l == nil || len(l.LayoutRanges) == 0 || !l.IsValid() {
+		return ""
+	}
+
+	// Binary search to find the first range whose end >= offset
+	left, right := 0, len(l.LayoutRanges)-1
+	for left < right {
+		mid := left + (right-left)/2
+		if l.LayoutRanges[mid].End < offset {
+			left = mid + 1
+		} else {
+			right = mid
+		}
+	}
+
+	// Range is guaranteed to exist, return its endpoint
+	return l.LayoutRanges[left].Endpoint
 }
